# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway of Restaurant Reservation System",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/reservation-system/reservations": {
            "get": {
                "description": "Retrieves multiple reservations info from reservations table in PostgreSQL",
                "tags": [
                    "reservation"
                ],
                "summary": "Fetches reservations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Number of reservations to fetch",
                        "name": "limit",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Number of reservations to omit",
                        "name": "offset",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.Reservations"
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while fetching reservations",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts new reservation info to reservations table in PostgreSQL",
                "tags": [
                    "reservation"
                ],
                "summary": "Creates a reservation",
                "parameters": [
                    {
                        "description": "New data",
                        "name": "new_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.ID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while creating reservation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/reservations/{reservation_id}": {
            "get": {
                "description": "Retrieves reservation info from reservations table in PostgreSQL",
                "tags": [
                    "reservation"
                ],
                "summary": "Gets a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while getting reservation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates reservation info in reservations table in PostgreSQL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Updates a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New info",
                        "name": "new_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation ID or data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while updating reservation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes reservation info from reservations table in PostgreSQL",
                "tags": [
                    "reservation"
                ],
                "summary": "Deletes a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while deleting reservation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/reservation-system/{reservation_id}/check": {
            "get": {
                "description": "Checks whether a reservation exists in reservations table in PostgreSQL",
                "tags": [
                    "reservation"
                ],
                "summary": "Validates a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.Status"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while checking reservation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/{reservation_id}/order": {
            "post": {
                "description": "Inserts order for a reservation in Redis",
                "tags": [
                    "reservation"
                ],
                "summary": "Orders meals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationOrders"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.ID"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation ID or data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while ordering",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/{reservation_id}/payment": {
            "post": {
                "description": "Inserts payment info to payments table in PostgreSQL",
                "tags": [
                    "reservation"
                ],
                "summary": "Pays for a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.Status"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while making a payment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "reservation.ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "reservation.ReservationDetails": {
            "type": "object",
            "properties": {
                "reservation_time": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "reservation.ReservationInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "reservation_time": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "reservation.ReservationOrder": {
            "type": "object",
            "properties": {
                "menu_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "reservation.ReservationOrders": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reservation.ReservationOrder"
                    }
                }
            }
        },
        "reservation.Reservations": {
            "type": "object",
            "properties": {
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reservation.ReservationInfo"
                    }
                }
            }
        },
        "reservation.Status": {
            "type": "object",
            "properties": {
                "successful": {
                    "type": "boolean"
                }
            }
        },
...
    }
}